another detection fix for bad_ftrace_value.py
i found this by playing with writing randomish values to ftrace_enabled,
and saw that this worked.

took a bit of effort to reproduce in python as i couldn't work out why
writing with a redirect had different behaviour than just f.write()

seems it has an extra write() with two null bytes.

Signed-off-by: B Horn <b@horn.uk>